Refactor #setSelectedBoundaryType
[#179577787]
