+++ js/stores/helpers/boundary-utils.ts
@@ -1,6 +1,6 @@
-import { toSpherical } from "../../geo-utils";
+import { toCartesian, toSpherical, getNorthVector } from "../../geo-utils";
 import getGrid from "../../plates-model/grid";
-import { BoundaryType, IBoundaryInfo } from "../../types";
+import { IBoundaryInfo, IHotSpot } from "../../types";
 import FieldStore from "../field-store";
 import ModelStore from "../model-store";
 
@@ -108,6 +108,60 @@ export function getBoundaryInfo(field: FieldStore, otherField: FieldStore): IBou
               : [field, otherField];
   }
   return { orientation, fields };
+}
+
+export function convertBoundaryTypeToHotSpots(boundaryInfo: IBoundaryInfo): IHotSpot[] {
+  const [field0, field1] = boundaryInfo.fields || [];
+  const type = boundaryInfo.type;
+  switch (boundaryInfo.orientation) {
+  case "northern-latitudinal":
+    if (field0) {
+      const latLon = toSpherical(field0.absolutePos);
+      const lat = latLon.lat + 0.15;
+      const lon = latLon.lon;
+      const newPos = toCartesian([lat, lon]);
+      const rotationAxis = newPos.clone().normalize();
+      const trueNorth = getNorthVector(lat, lon);
+      const force = type === "convergent" ? trueNorth.applyAxisAngle(rotationAxis, Math.PI) : trueNorth;
+      return [{ position: newPos, force }];
+    }
+    break;
+  case "longitudinal":
+    if (field0 && field1) {
+      const latLon0 = toSpherical(field0.absolutePos);
+      const lat0 = latLon0.lat;
+      // TODO: it should be adjusted for extreme lat values in a better way
+      const kLonDiff = 0.2 + Math.pow(Math.abs(lat0), 2) * 0.2;
+      const lon0 = latLon0.lon - kLonDiff;
+      const latLon1 = toSpherical(field1.absolutePos);
+      const lat1 = latLon1.lat;
+      const lon1 = latLon1.lon + kLonDiff;
+      const latAvg = 0.5 * (lat0 + lat1);
+      const newPos0 = toCartesian([latAvg, lon0]);
+      const newPos1 = toCartesian([latAvg, lon1]);
+      const rotationAxis0 = newPos0.clone().normalize();
+      const rotationAxis1 = newPos1.clone().normalize();
+      const trueNorth0 = getNorthVector(latAvg, lon0);
+      const trueNorth1 = getNorthVector(latAvg, lon1);
+      const force0 = type === "convergent" ? trueNorth0.applyAxisAngle(rotationAxis0, 0.5 * Math.PI) : trueNorth0.applyAxisAngle(rotationAxis0, -0.5 * Math.PI);
+      const force1 = type === "convergent" ? trueNorth1.applyAxisAngle(rotationAxis1, -0.5 * Math.PI) : trueNorth1.applyAxisAngle(rotationAxis1, 0.5 * Math.PI);
+      return [{ position: newPos0, force: force0 }, { position: newPos1, force: force1 }];
+    }
+    break;
+  case "southern-latitudinal":
+    if (field1) {
+      const latLon = toSpherical(field1.absolutePos);
+      const lat = latLon.lat - 0.15;
+      const lon = latLon.lon;
+      const newPos = toCartesian([lat, lon]);
+      const rotationAxis = newPos.clone().normalize();
+      const trueNorth = getNorthVector(lat, lon);
+      const force = type === "convergent" ? trueNorth : trueNorth.applyAxisAngle(rotationAxis, Math.PI);
+      return [{ position: newPos, force }];
+    }
+    break;
+  }
+  return [];
 }
 
 export function findBoundaryFieldAround(field: FieldStore | null, model: ModelStore) {
